Fix flaky test for ExifInterfaceTest
ExifInterface#setDateTime uses the last 3 digits of the given
timestamp to create the value for SubSecTime tag. However, if the
value is less than 3 digits (e.g. 89 or 6), it incorrectly added
zeroes at the end to make the value 3 digits. This CL prevents this
by adding zeroes to the front of the value (e.g. "089" or "006")
instead of at the end.

Bug: 174553127
Test: Run testSubsec 10 times
Change-Id: I4b24c658a730146c8a244db15347373264900a39