Fix nullability lint issue with ExifInterface
This change only adds clarity around the nullability of the `location` param on `setGpsInfo` and indicates that the method is a No-Op if the param is `null`.

Ideally setting `location` to `null` would also clear related GPS values, but that is beyond the scope of this fix.

Relnote: "Indicate that the `location` param for `setGpsInfo` is `@Nullable` and that passing `null` will result in a No-Op."

Fixes: 236484611

Test: ./gradlew :exifinterface:exifinterface:lintDebug
Test: presubmit
Change-Id: If924cd23c58f32cbb112db98cdd3d988105c9429