ExifInterface: Write new XMP data out to a separate segment
The XMP spec part 3 section 3.3.2 makes it clear that JPEG files should
store XMP data in separate APP1 segment, not the Exif segment under tag
700. ExifInterface is documented to (incorrectly) prefer XMP data stored
in the Exif segment if both are present. But the behaviour of where XMP
data is added in the file if it wasn't previously present isn't
explicitly documented. This change ensures that XMP data added to a file
that didn't previously have any is added in a standalone segment.

This change also adds some additional documentation to describe this
behaviour.

Test: ExifInterfaceTest
Change-Id: I84328802a3b43b4766850fcdc6731c1e7c59e909