Fix handling of null XMP data
Passing a null value to ExifInterface.setAttribute(..) is expected
to clear that attribute.

When passing null with TAG_XMP, ExifInterface was throwing a
NullPointerException. This change adds a test for this case, and fixes
the NPE. It also fixes the handling of XMP in PNG files to ensure that
the XMP is correctly removed in this case.

This change also tightens the handling of iTXt PNG chunks when writing
XMP data, to avoid accidentally overwriting a non-XMP chunk.

Bug: 405676710
Test: ExifInterfaceTest
Change-Id: I34a9de734d78d539ee841545c23df67444298a0c